refactor(tarcap): tarcap source code folders/files naming refactor
