+++ src/network/tarcap/new_taraxa_capability.cpp
@@ -7,23 +7,23 @@
 #include "consensus/vote.hpp"
 #include "dag/dag.hpp"
 #include "network/tarcap/node_stats.hpp"
-#include "network/tarcap/packets_handler/handlers/blocks_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/common/syncing_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/dag_packets_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/get_blocks_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/get_pbft_block_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/new_pbft_block_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/pbft_block_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/status_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/test_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/transaction_packet_handler.hpp"
-#include "network/tarcap/packets_handler/handlers/vote_packets_handler.hpp"
-#include "network/tarcap/packets_handler/packets_handler.hpp"
-#include "network/tarcap/packets_handler/peers_state.hpp"
-#include "network/tarcap/packets_handler/syncing_state.hpp"
-#include "network/tarcap/packets_handler/test_state.hpp"
-#include "network/tarcap/packets_handler/taraxa_peer.hpp"
+#include "network/tarcap/packets_handlers/blocks_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/common/syncing_handler.hpp"
+#include "network/tarcap/packets_handlers/dag_packets_handler.hpp"
+#include "network/tarcap/packets_handlers/get_blocks_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/get_pbft_block_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/new_pbft_block_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/pbft_block_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/status_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/test_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/transaction_packet_handler.hpp"
+#include "network/tarcap/packets_handlers/vote_packets_handler.hpp"
+#include "network/tarcap/shared_states/peers_state.hpp"
+#include "network/tarcap/shared_states/syncing_state.hpp"
+#include "network/tarcap/shared_states/test_state.hpp"
 #include "node/full_node.hpp"
+#include "packets_handler.hpp"
+#include "taraxa_peer.hpp"
 #include "transaction_manager/transaction_manager.hpp"
 
 namespace taraxa::network::tarcap {
