+++ src/network/tarcap/packets_handlers/get_blocks_packet_handler.cpp
@@ -1,8 +1,8 @@
 #include "get_blocks_packet_handler.hpp"
 
 #include "dag/dag.hpp"
-#include "network/tarcap/packets_handler/handlers/common/get_blocks_request_type.hpp"
-#include "network/tarcap/packets_handler/syncing_state.hpp"
+#include "network/tarcap/packets_handlers/common/get_blocks_request_type.hpp"
+#include "network/tarcap/shared_states/syncing_state.hpp"
 #include "transaction_manager/transaction_manager.hpp"
 
 namespace taraxa::network::tarcap {
