+++ src/network/tarcap/packets_handlers/test_packet_handler.hpp
@@ -4,7 +4,7 @@
 
 #include <mutex>
 
-#include "network/tarcap/packets_handler/handlers/common/packet_handler.hpp"
+#include "network/tarcap/packets_handlers/common/packet_handler.hpp"
 
 namespace taraxa::network::tarcap {
