+++ src/network/tarcap/packets_handlers/transaction_packet_handler.cpp
@@ -1,7 +1,7 @@
 #include "transaction_packet_handler.hpp"
 
 #include "dag/dag_block_manager.hpp"
-#include "network/tarcap/packets_handler/test_state.hpp"
+#include "network/tarcap/shared_states/test_state.hpp"
 #include "transaction_manager/transaction_manager.hpp"
 
 namespace taraxa::network::tarcap {
